changed c#/python communication to be synchronous
mono seems to be broken when to many async reads are used. (Same behaviour with pipes and sockets)
synchronous communication makes it more difficult to react if the evaluation script (python) does not respond. Currently no error handling.
Asynchronous communication should(!!! not well tested yet) not be necessary any more.